Removed some classes, added some new classes
For simplicity, Combat Card is removed. A Special Card doesn't have
combat stats, so it's not a combat card. Only the Group Card and
Illuminati Card have combat stats. However the Combat Card class is
confusing for the team, so it will be removed to hopefully reduce
confusions.

Group class renamed to GroupCard

Card class
- Added 3 constructors for Illuminati Card, Group Card, and Special Card
- Type 1: Illuminati
- Type 2: Group Card
- Type 3: Special Card

Arrow class, IncomingArrow class, and Outgoing Arrow class
- The Arrow class applies to Group Cards and Illuminati Cards only.
Special Cards do not have arrows. If there is a better way to represent
arrows, the refactorings/suggestions are welcomed.

Game class
- Contains the game logic, based on the flowchart