Can play a menu game. Still have more work to do.
Make sure all java files are under the package theilluminatigame;
Run main from Game.java
In the console, keep spamming 2's and you will get to the Attack to
Destroy menu, as well as a test scenario of an attack to destroy.
A challenge is the Attack to Control feature because of the arrows.